quic: set max_[stream_]window

+++ quic/quic-server.cpp
@@ -146,10 +146,13 @@ void QuicServer::log_stats(std::string reason) {
 void QuicServer::log_conn_stats(ConnectionState &state, const char *reason) {
   constexpr double kNsToMs = 1e-6;
   auto info = state.impl().get_conn_info();
-  double loss_pct = info.pkt_sent ? (100.0 * static_cast<double>(info.pkt_lost) / info.pkt_sent) : 0.0;
+  double loss_pct =
+      info.pkt_sent ? (100.0 * static_cast<double>(info.pkt_lost) / static_cast<double>(info.pkt_sent)) : 0.0;
   LOG(INFO) << "quic stats (" << reason << ") for " << state.remote_address << " cid=" << state.cid
-            << " rtt_ms{smoothed=" << info.smoothed_rtt * kNsToMs << " min=" << info.min_rtt * kNsToMs
-            << " latest=" << info.latest_rtt * kNsToMs << " var=" << info.rttvar * kNsToMs << "}"
+            << " rtt_ms{smoothed=" << static_cast<double>(info.smoothed_rtt) * kNsToMs
+            << " min=" << static_cast<double>(info.min_rtt) * kNsToMs
+            << " latest=" << static_cast<double>(info.latest_rtt) * kNsToMs
+            << " var=" << static_cast<double>(info.rttvar) * kNsToMs << "}"
             << " cwnd=" << info.cwnd << " inflight=" << info.bytes_in_flight << " sent=" << info.pkt_sent << "/"
             << info.bytes_sent << " recv=" << info.pkt_recv << "/" << info.bytes_recv << " lost=" << info.pkt_lost
             << "/" << info.bytes_lost << " loss=" << loss_pct << "%";
